+++ src/Stack/Config.hs
@@ -64,6 +64,7 @@ import           Stack.Config.Docker
 import           Stack.Config.Nix
 import           Stack.Constants
 import           Stack.Build.Haddock (shouldHaddockDeps)
+import           Stack.PersistentTH (initCacheStorage)
 import           Stack.SourceMap
 import           Stack.Types.Build
 import           Stack.Types.Config
@@ -359,7 +360,9 @@ configFromConfigMonoid
        hsc
        (maybe HpackBundled HpackCommand $ getFirst configMonoidOverrideHpack)
        clConnectionCount
-       (\configPantryConfig -> inner Config {..})
+       (\configPantryConfig -> initCacheStorage
+         (configStackRoot </> relFileCaches)
+         (\configCachePool -> inner Config {..}))
 
 -- | Get the default location of the local programs directory.
 getDefaultLocalProgramsBase :: MonadThrow m